+++ kernel/scheduler.c
@@ -14,7 +14,7 @@ static uint32_t sched_tick     = 0;
 
 static int strncmp_k(const char *a, const char *b, int n) {
     while (n-- > 0) {
-        if (*a != *b) return (unsigned char)*a - (unsigned char)*b;
+        if (*a != *b) return (int)(unsigned char)*a - (int)(unsigned char)*b;
         if (*a == '\0') return 0;
         a++; b++;
     }
